Renaming the Cassini model to HDF5, updating documentation.
