+++ requirements.txt
@@ -9,7 +9,6 @@ kiwisolver==1.3.2
 matplotlib==3.4.2
 numpy==1.20.2
 packaging==21.3
-pandas==1.3.4
 Pillow==8.4.0
 pyerfa==2.0.0.1
 pyparsing==3.0.4
